feat(app): bit of refactoring, seperated Anime page into smaller componenents and finished the details component except for recommendations
