+++ client/src/components/Anime/SideInfo.tsx
@@ -1,0 +1,111 @@
+import { VStack, Image, Text, Flex, Spinner, Alert, AlertIcon, AlertTitle } from '@chakra-ui/react';
+import _ from 'lodash';
+import React from 'react';
+import { useQuery } from 'react-query';
+import { useParams } from 'react-router-dom';
+import { combinedAnimeResponse } from '../../api/api';
+
+export const SideInfo = () => {
+  const { id } = useParams<{ id: string }>();
+  const parsedId = parseInt(id);
+
+  const { isLoading, isFetching, data, error, isError } = useQuery(
+    'anime',
+    () => {
+      return combinedAnimeResponse('anime', parsedId);
+    },
+    { refetchOnWindowFocus: false }
+  );
+
+  // if (isLoading || isFetching) {
+  //   return (
+  //     <Flex h="100vh" justify="center" mt={50}>
+  //       <Spinner color="blue" />
+  //     </Flex>
+  //   );
+  // }
+
+  if (isError) {
+    <Alert status="error">
+      <AlertIcon />
+      <AlertTitle mr={2}>{error}placeholder</AlertTitle>
+    </Alert>;
+  }
+
+  if (!data) {
+    return null;
+  }
+
+  const { details } = data;
+
+  const studios = _.map(details.studios, (studio) => {
+    return studio.name;
+  });
+
+  const producers = _.map(details.producers, (producer) => {
+    return producer.name;
+  });
+
+  const genres = _.map(details.genres, (genre) => {
+    return genre.name;
+  });
+  return (
+    <VStack w={300} px={3} pt={2} align="left" fontSize="sm">
+      <Image w={275} fit="cover" src={details.image_url} />
+      <Text borderBottom="1px solid black" fontWeight="bold" pt={5}>
+        Alternative Titles
+      </Text>
+      <Text>
+        <b>English:</b> {details.title_english}
+      </Text>
+      <Text>
+        <b>Synonym:</b> {details.title_synonyms.join(', ')}
+      </Text>
+      <Text>
+        <b>Japanese:</b> {details.title_japanese}
+      </Text>
+      <Text borderBottom="1px solid black" fontWeight="bold" pt={5}>
+        Information
+      </Text>
+      <Text>
+        <b>Type:</b> {details.type}
+      </Text>
+      <Text>
+        <b>Episodes:</b> {details.episodes}
+      </Text>
+      <Text>
+        <b>Status:</b> {details.status}
+      </Text>
+      <Text>
+        <b>Premiered:</b> {details.premiered}
+      </Text>
+      <Text>
+        <b>Producers:</b> {producers.join(', ')}
+      </Text>
+      <Text>
+        <b>Studios:</b> {studios.join(', ')}
+      </Text>
+      <Text>
+        <b>Sources:</b> {details.source}
+      </Text>
+      <Text>
+        <b>Genres:</b> {genres.join(', ')}
+      </Text>
+      <Text>
+        <b>Duration:</b> {details.duration}
+      </Text>
+      <Text>
+        <b>Rating:</b> {details.rating}
+      </Text>
+      <Text borderBottom="1px solid black" fontWeight="bold" pt={5}>
+        Statistics
+      </Text>
+      <Text>
+        <b>Score:</b> {details.score}
+      </Text>
+      <Text>
+        <b>Ranked:</b> {details.rank}
+      </Text>
+    </VStack>
+  );
+};
